% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun1_1_import_data_FEA.R
\name{fun1_1_import_data_FEA}
\alias{fun1_1_import_data_FEA}
\title{Import data locally (and convert it to a tibble for convinience - easier to work with in the console than a data.frame)}
\usage{
fun1_1_import_data_FEA()
}
\value{
a tibble with the data (invisible)
}
\description{
FEA refers to "forlopsoversikt", "enkeltledd" and "alle scorer"
}
\details{
Note: This function's name when writing aarsrapport 2019 (in 2020) was: import11_data)

Use this function like this:
RegData <- fun1_1_import_data_FEA()
}
