% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun2_1_1_RegData_newVarGlobal.R
\name{fun2_1_1_RegData_newVarGlobal}
\alias{fun2_1_1_RegData_newVarGlobal}
\title{Make variable with hospital department names (Global refers to that the varaibel(s) made her will be used globally/often)}
\usage{
fun2_1_1_RegData_newVarGlobal(myInData)
}
\arguments{
\item{myInData}{RegData

Output is data, with new var. that will be used extensively (like AvdNavn which is name of hospital unit)

Note: This function's name when writing aarsrapport 2019 (in 2020) was: make21_data_newVarGlobal}
}
\value{
RegData (invisible)
}
\description{
Make variable with hospital department names (Global refers to that the varaibel(s) made her will be used globally/often)
}
