% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun2_1_2_RegData_newVarFrmt.R
\name{fun2_1_2_RegData_newVarFrmt}
\alias{fun2_1_2_RegData_newVarFrmt}
\title{Make variables with new formats from old variables (Frmt refers to formatted)}
\usage{
fun2_1_2_RegData_newVarFrmt(myInData)
}
\arguments{
\item{myInData}{RegData}
}
\value{
RegData (invisible)
}
\description{
Output is data, with new var. which is similar to old, but with new format and "FRMT" suffix.
}
\details{
Note: This function's name when writing aarsrapport 2019 (in 2020) was: make22_data_newVarFrmt
}
