% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun2_1_3_RegData_newVarMAsNA.R
\name{fun2_1_3_RegData_newVarMAsNA}
\alias{fun2_1_3_RegData_newVarMAsNA}
\title{Make variables where missing is coded as NA (MAsNA)}
\usage{
fun2_1_3_RegData_newVarMAsNA(myInData)
}
\arguments{
\item{myInData}{RegData}
}
\value{
myInData (invisble)
}
\description{
Output is data, with new variables where missing is coded as NA. Some are categoricial, and som have values that are labels.
Accordingly the new variables are named with suffixes _MISSING, CAT, and _NAMES (TODO: rename to for instance  MaSNA, NumCat and LabCat)
}
\details{
Note: This function's name when writing aarsrapport 2019 (in 2020) was: make23_data_newVar
}
