% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun2_1_4_RegData_newVarDich.R
\name{fun2_1_4_RegData_newVarDich}
\alias{fun2_1_4_RegData_newVarDich}
\title{Make dichotomous variables (dich)}
\usage{
fun2_1_4_RegData_newVarDich(myInData)
}
\arguments{
\item{myInData}{RegData}
}
\value{
myInData (invisible)
}
\description{
Output is data, with new variables that ar dichotomous.
}
\details{
Note: This function's name when writing aarsrapport 2019 (in 2020) was: make24_data_newVar01
}
