% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun2_2_RegDataBeh_newVarGlobal.R
\name{fun2_2_RegDataBeh_newVarGlobal}
\alias{fun2_2_RegDataBeh_newVarGlobal}
\title{Make RegDatabeh with new global variables (Global refers to that the varaibel(s) made her will be used globally/often)}
\usage{
fun2_2_RegDataBeh_newVarGlobal(myInData1 = RegData,
  myInData2 = RegDataBeh)
}
\arguments{
\item{myInData1}{RegData}

\item{myInData2}{RegDataBeh}
}
\description{
fun2_1_1.. must be run first.
}
