% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun2_4_1_RegDataStartEnd.R
\name{fun2_4_1_RegDataStartEnd}
\alias{fun2_4_1_RegDataStartEnd}
\title{Make start-end data:}
\usage{
fun2_4_1_RegDataStartEnd(myInData)
}
\arguments{
\item{myInData}{RegData}
}
\value{
RegDataStartEnd
}
\description{
* We still want one line per patient, and thus move some of the variables from the final registration to the patients start registration.
* The End registrations will appear twice in this data. Once as in the normal data and in addition added to the start registrations as new variables
}
\details{
Note: This function's name when writing aarsrapport 2019 (in 2020) was: make26_dataStartEnd()
}
