% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun2_4_2_RegDataStartEnd_newVar.R
\name{fun2_4_2_RegDataStartEnd_newVar}
\alias{fun2_4_2_RegDataStartEnd_newVar}
\title{Title}
\usage{
fun2_4_2_RegDataStartEnd_newVar(myInData = RegDataStartEnd)
}
\arguments{
\item{myInData}{RegDataStartEnd}
}
\value{
RegDataStartEnd (invisible)
}
\description{
Make new variables in start-end data
Quality indicators
* EDEQ60GlobalScore_CHANGE_PROP
*...
}
\details{
Data quality variables or the quality indicators
*...
* MedBMI_missEnd_whenU185Start

Note: This function's name when writing aarsrapport 2019 (in 2020) was: make27_dataStartEnd_NewVar
}
