% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun2_5_RegDataStartEndNatVal.R
\name{fun2_5_RegDataStartEndNatVal}
\alias{fun2_5_RegDataStartEndNatVal}
\title{Make start-end data with national values#'}
\usage{
fun2_5_RegDataStartEndNatVal(myInData)
}
\arguments{
\item{myInData}{RegDataStartEnd}
}
\value{
RegDataStartEnd_NatVal
}
\description{
Data, start-end, with national values, double the size of the start-end data, where half will represent national data,
}
\details{
Note: This function's name when writing aarsrapport 2019 (in 2020) was: make28_dataStartEnd_NatVal
}
