% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun2_dataList.R
\name{fun2_dataList}
\alias{fun2_dataList}
\title{List all the prepared datasets for norspis2 package}
\usage{
fun2_dataList(myInData1 = RegData, myInData2 = RegDataBeh)
}
\arguments{
\item{myInData1}{RegData}

\item{myInData2}{RegDataBeh}
}
\value{
dataList, a list with all five datasets
}
\description{
This function runs the functions fun2_1_1.. to fun2_1_5 in the correct order, to make a list of
five datasets (the "2" suffix refers to that these data can be used to make figures and tables in the norspis2 package):
}
\details{
* RegData2
* RegDataBeh2
* RegDataNatVal2
* RegDataStartEnd2
* RegDataStartEndNatVal2
}
