% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun3_1_filter_RegData.R
\name{fun3_1_filter_RegData}
\alias{fun3_1_filter_RegData}
\title{Filter function for RegData (NOT start-end data):}
\usage{
fun3_1_filter_RegData(
  RegData = myInData,
  regStatus = c(0, 1, -1),
  regType = c(0, 1, 2, 3, 4, 5, 6, 98, 99),
  dateFrom = "2012-01-01",
  dateTo = "2100-12-31",
  ageFrom = 0,
  ageTo = 200,
  BUP = 99,
  AvdAlder = 99
)
}
\arguments{
\item{RegData}{myInData}

\item{regStatus}{c(0,1,-1)}

\item{regType}{c(0,1,2,3,4,5,6,98,99)}

\item{dateFrom}{"2012-01-01",}

\item{dateTo}{"2100-12-31"}

\item{ageFrom}{0}

\item{ageTo}{200}
}
\value{
RegData_filtered
}
\description{
This function is also used to filter RegData with national values (NatVal) (as long as they are NOT start-end data)
}
\details{
Note: This function's name when writing aarsrapport 2019 (in 2020) was: make29_data_filtered
}
