% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun3_2_filter_RegDataStartEnd.R
\name{fun3_2_filter_RegDataStartEnd}
\alias{fun3_2_filter_RegDataStartEnd}
\title{Filter function for start-end registrations (RegDataStartEnd):}
\usage{
fun3_2_filter_RegDataStartEnd(RegDataStartEnd = myInData,
  BasisRegStatus.x = c(0, 1, -1), BasisRegStatus.y = c(0, 1, -1, NA),
  RegRegType.x = c(0, 1, 2, 3, 4, 5, 6, 98, 99), RegRegType.y = c(5, 6,
  98, 99, NA), dateFrom.x = "2012-01-01", dateTo.x = "2100-12-31",
  dateFrom.y = "2012-01-01", dateTo.y = "2100-12-31", ageFrom.x = 0,
  ageTo.x = 200, ageFrom.y = 0, ageTo.y = 200)
}
\arguments{
\item{RegDataStartEnd}{myInData}

\item{BasisRegStatus.x}{c(0,1,-1)}

\item{RegRegType.x}{c(0,1,2,3,4,5,6,98,99)}

\item{RegRegType.y}{c(5,6,98,99,NA)}

\item{dateFrom.x}{"2012-01-01"}

\item{dateTo.x}{"2100-12-31"}

\item{dateFrom.y}{"2012-01-01"}

\item{dateTo.y}{"2100-12-31"}

\item{ageFrom.x}{0}

\item{ageTo.x}{200}

\item{ageFrom.y}{0}

\item{ageTo.y}{200}
}
\value{
RegData_filtered
}
\description{
We use a separate filter function for start-end data. Star-data have the suffix x, and end-data have the suffix y.
We can thus for instance filter on dateFrom.x and dateTo.x, if we want to filter on dates of the start registrations,
or dateFrom.y and dateTo.y if we want to filter on dates of end registrations.
}
\details{
Note: This function's name when writing aarsrapport 2019 (in 2020) was: make291_data_filtered
}
