% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_figFig_variableDist.R
\name{make_figFig_variableDist}
\alias{make_figFig_variableDist}
\title{Function to make simple figure with distribution of a variable's values (nationally)}
\usage{
make_figFig_variableDist(
  my_proptable,
  mytitle,
  comparison = FALSE,
  my_proptable_compare
)
}
\arguments{
\item{my_proptable}{table from make_figTable_variableCompar()}

\item{mytitle}{title of plot}
}
\value{
figure
}
\description{
Note: This function's name when writing aarsrapport 2019 (in 2020) was: make_propfig
}
