% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_figTable_variableDist.R
\name{make_figTable_variableDist}
\alias{make_figTable_variableDist}
\title{Function to make data table to go into simple plots of distributions of variables}
\usage{
make_figTable_variableDist(myvar, mycat, mylab, myInData)
}
\arguments{
\item{myvar}{e.g. expr(PO10Pasientsikkerhet)}

\item{mycat}{e.g. expr(c(0,1,2,3,4))}

\item{mylab}{e.g. vars("0"="Ikke i det hele tatt", 1"="I liten grad", "2"="I noen grad", "3"="I stor grad","4"="I svært stor grad")}

\item{myInData}{myInData <- RegDataStartEnd}
}
\value{
summary table with propoprtions of answers withing each variable category, to go into make_figFig_variableDist
}
\description{
Note: This function's name when writing aarsrapport 2019 (in 2020) was: make_proptable
}
