% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_table_DQ_completeness_endreg.R
\name{make_table_DQ_completeness_endreg}
\alias{make_table_DQ_completeness_endreg}
\title{Title}
\usage{
make_table_DQ_completeness_endreg(
  RegDataStartEnd = DL$RegDataStartEnd2,
  userRole,
  reshID,
  dateFrom = "2012-01-01",
  dateTo = "2021-12-31",
  timeseries = F
)
}
\arguments{
\item{timeseries}{}
}
\value{

}
\description{
Make data quality table with completeness of registrations (proportion of
start registrations that has delivered end registration("forlopskompletthet)).
The measure is based on expected number of delivered end registrations,
Expected number of delivered end registrations is calculated by use of
historical treatment times given by end registrations that actually have been
delivered.
}
\details{
The method used is outlines in NorSpis' annual report for 2019 and 2020 at
published at www.kvalitetsregistre.no.

Since we use "start.end" data, we identify each individual start
registration that has an end registration, when we make the counts.

All registrations since start of NorSpis is included in the calculation of
treatment time and empirical cumulative distribution function.
}
