% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_table_DQ_missing.R
\name{make_table_DQ_missing}
\alias{make_table_DQ_missing}
\title{Make data quality table with proportions of missing values for chosen variables.}
\usage{
make_table_DQ_missing(RegData = myFilteredData,
  varsInMissTab = myvarStringMiss)
}
\arguments{
\item{RegData}{myFilteredData}

\item{varsInMissTab}{myvarStringMiss <- myvarStringMiss <- c(quo(variablename_miss), quo(variablename_missStart))}
}
\value{
output_missingTibble
}
\description{
Note: This function's name when writing aarsrapport 2019 (in 2020) was: make_missingTab
}
