% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_table_DQ_regCount_asFt.R
\name{make_table_DQ_regCount_asFt}
\alias{make_table_DQ_regCount_asFt}
\title{Make table of count of registrations into a nicely formatted Flextable}
\usage{
make_table_DQ_regCount_asFt(tab)
}
\arguments{
\item{tab}{this is the table made with the function make_table_DQ_regCount}
}
\description{
This functionality did not have its own function in aarsrapport 2019
(in 2020) but the code was included in the .Rmd file for the
"aarsrapport"
}
