% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_table_DQ_regCountYearly.R
\name{make_table_DQ_regCountYearly}
\alias{make_table_DQ_regCountYearly}
\title{Make table with number of registrations per department per year}
\usage{
make_table_DQ_regCountYearly(
  RegDataNatValFiltered,
  saveAsImage = F,
  saveAsDoc = F,
  pathToSaveTableFile = "F:/",
  fileNameSuffix = "_2021"
)
}
\arguments{
\item{RegDataNatVal}{data with own rows representing national values}
}
\value{
flexable of number of registrations per year per department
}
\description{
The data must be filtered first, to contain the registration types
you want to present, if not the table will show a mix of registration types.
}
