% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_table_patChar.R
\name{make_table_patChar}
\alias{make_table_patChar}
\title{Make table with overview of patient characteristics:}
\usage{
make_table_patChar(RegData = myFilteredData, varsInTab = myvarString)
}
\arguments{
\item{RegData}{myFilteredData}

\item{varsInTab}{myvarString <- c(quo(variablename_MISSING_NAMES), quo(variablename_MISSING_NAMES))}
}
\value{
output_tibble
}
\description{
Note: This function's name when writing aarsrapport 2019 (in 2020) was: make_popCharTab
}
