% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_table_popchar.R
\name{make_table_popchar}
\alias{make_table_popchar}
\title{make_table_popchar}
\usage{
make_table_popchar(
  RegData = myFilteredData,
  varsInTab = myvarString,
  varsNames = myvarNames,
  comparison = FALSE,
  dont_visualize_nei = FALSE,
  dont_visualize_mangler = FALSE,
  mergevalues_2nd_column = TRUE,
  row_height_ft = 0.25,
  font_size_ft = 8,
  lollipop_max_ft = 10,
  lollipop_min_ft = -10
)
}
\arguments{
\item{font_size_ft}{}
}
\value{

}
\description{
Function for table with overview of patient characteristics.
}
\details{
The name of this function when maing aarsrapport 2019, in 2020,  was make_popCharTab().
}
