% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_filter_attributedata.R
\name{map_filter_attributedata}
\alias{map_filter_attributedata}
\title{Function that filter the data (using dplyr syntax)
An own function because we want to use the returned data several places later
(in both plot and table).}
\usage{
map_filter_attributedata(
  chosen_unit_type = c(1, 2, 4, 5, 6, 7, 8, 9, 3, 10, 11, 12),
  chosen_hf = "Alle"
)
}
\arguments{
\item{chosen_unit_type}{1,2,4,5,6,7,8,9,3,10}

\item{chosen_map_area}{norway, north, middle, west, southeast}
}
\value{
attributedata_geo_chosenarea Prepared dataset with attributes and geometry column *for the chosen area*
}
\description{
Function that filter the data (using dplyr syntax)
An own function because we want to use the returned data several places later
(in both plot and table).
}
