% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_make_backmap.R
\name{map_make_backmap}
\alias{map_make_backmap}
\title{Function that makes the background map.}
\usage{
map_make_backmap(
  chosen_hf = c("Alle", "Helse Nord", "Helse Midt", "Helse Vest", "Helse Sør-Øst",
    "Helse Møre og Romsdal", "Helse Nord-Trøndelag", "St. Olavs Hospital",
    "Finnmarkssykehuset", "Helgelandssykehuset", "Nordlandssykehuset",
    "Universitetssykehuset Nord-Norge", "Private behandlingssteder",
    "Akershus universitetssykehus", "Oslo universitetssykehus", "Sykehuset i Vestfold",
    "Sykehuset Innlandet", "Sykehuset Telemark", "Sykehuset Østfold",
    "Sørlandet Sykehus", "Vestre Viken", "Helse Bergen", "Helse Fonna", "Helse Førde", 
        "Helse Stavanger")
)
}
\arguments{
\item{chosen_map_area}{norway, north, middle, west, southeast}
}
\value{
backgroundmap
}
\description{
Function that makes the background map.
}
