% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_make_map.R
\name{map_make_map}
\alias{map_make_map}
\title{Function that makes the the final map.(map_make_map uses the function map_make_backmap)}
\usage{
map_make_map(
  chosen_unit_type = c(1, 2, 4, 5, 6, 7, 8, 9, 3, 10, 11, 12),
  chosen_hf = c("Alle", "Helse Nord", "Helse Midt", "Helse Vest", "Helse Sør-Øst",
    "Helse Møre og Romsdal", "Helse Nord-Trøndelag", "St. Olavs Hospital",
    "Finnmarkssykehuset", "Helgelandssykehuset", "Nordlandssykehuset",
    "Universitetssykehuset Nord-Norge", "Private behandlingssteder",
    "Akershus universitetssykehus", "Oslo universitetssykehus", "Sykehuset i Vestfold",
    "Sykehuset Innlandet", "Sykehuset Telemark", "Sykehuset Østfold",
    "Sørlandet Sykehus", "Vestre Viken", "Helse Bergen", "Helse Fonna", "Helse Førde", 
        "Helse Stavanger"),
  chosen_color_var = "egensk_tilslutning2019_hfavd_OK",
  chosen_size_var = "egensk_n_pop",
  chosen_shape_var = "enhet_kat_sf_num",
  visualize_points = "yes",
  visualize_text_unitname = "no",
  visualize_colour_polygons = "no",
  visualize_size_points = "no"
)
}
\arguments{
\item{chosen_unit_type}{c(1,2,4,5,6,7,8,9,3,10,11,12)}

\item{chosen_color_var}{'egensk_tilslutning2019_hfavd_OK'}

\item{chosen_size_var}{'egensk_n_pop'}

\item{chosen_shape_var}{'enhet_kat_sf_num'}

\item{visualize_text_unitname}{'yes'}

\item{visualize_colour_polygons}{'yes'}

\item{visualize_size_points}{'no'}

\item{chosen_map_area}{norway, north, middle, west, southeast}
}
\value{
Plots the map.
}
\description{
Function that makes the the final map.(map_make_map uses the function map_make_backmap)
}
