% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_make_table.R
\name{map_make_table}
\alias{map_make_table}
\title{Function that makes the table with map information pluss other variables of choice}
\usage{
map_make_table(
  chosen_unit_type = c(1, 2, 4, 5, 6, 7, 8, 9, 3, 10, 11, 12),
  chosen_hf = c("Alle", "Helse Nord", "Helse Midt", "Helse Vest", "Helse Sør-Øst",
    "Helse Møre og Romsdal", "Helse Nord-Trøndelag", "St. Olavs Hospital",
    "Finnmarkssykehuset", "Helgelandssykehuset", "Nordlandssykehuset",
    "Universitetssykehuset Nord-Norge", "Private behandlingssteder",
    "Akershus universitetssykehus", "Oslo universitetssykehus", "Sykehuset i Vestfold",
    "Sykehuset Innlandet", "Sykehuset Telemark", "Sykehuset Østfold",
    "Sørlandet Sykehus", "Vestre Viken", "Helse Bergen", "Helse Fonna", "Helse Førde", 
        "Helse Stavanger"),
  chosen_string_var = "egensk_datakval_kompl_dg2017_hfavd_OK",
  chosen_color_var = "egensk_tilslutning2019_hfavd_OK",
  chosen_color_var2 = "egensk_tilslutning2018_hfavd_OK",
  chosen_color_var3 = "null",
  chosen_table_type = "kable"
)
}
\arguments{
\item{chosen_unit_type}{write description of variable here}

\item{chosen_string_var}{write description of variable here}

\item{chosen_color_var}{write description of variable here}

\item{chosen_color_var2}{write description of variable here}

\item{chosen_color_var3}{write description of variable here}

\item{chosen_map_area}{write description of variable here}
}
\value{
Shows the table.
}
\description{
Function that makes the table with map information pluss other variables of choice
}
