% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NorSpis1_1_Preprosess.R
\name{NorSpis1_1_Preprosess}
\alias{NorSpis1_1_Preprosess}
\title{Preprosesser data fra NorSpis}
\usage{
NorSpis1_1_Preprosess(RegData = RegData)
}
\arguments{
\item{RegData}{}
}
\value{
Data En liste med det filtrerte datasettet (og sykehusnavnet som tilsvarer reshID, ikke pt)
}
\description{
Denne funksjonen gir nytt navn til variabler og beregner eventuelt nye.
}
