% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NorSpis1_2_VarTilrettelegg.R
\name{NorSpis1_2_VarTilrettelegg}
\alias{NorSpis1_2_VarTilrettelegg}
\title{Funksjon for å tilrettelegge variable for beregning.}
\usage{
NorSpis1_2_VarTilrettelegg(RegData, valgtVar, grVar = "SykehusNavn",
  figurtype = "andeler")
}
\arguments{
\item{figurtype}{}
}
\value{
Definisjon av valgt variabel.
}
\description{
Denne funksjonen gjør utvalg og tilrettelegger variable (gitt ved valgtVar) til videre bruk.
Videre bruk kan eksempelvis være beregning av AggVerdier eller gjennomsnitt.
Funksjonen gjør også filtreringer som å fjerne ugyldige verdier for den valgte variabelen, samt ta høyde for avhengigheter med
andre variable. Det er også her man angir aksetekster og titler for den valgte variabelen.
Her kan mye hentes til analysebok
}
