% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NorSpis1FigPrePost.R
\name{NorSpis1FigPrePost}
\alias{NorSpis1FigPrePost}
\title{Søylediagram som viser sentralmå, gjennomsnitt/median, ved to ulike tidspunkter
(før og etter/ved starten og slutten av behandlingen)}
\usage{
NorSpis1FigPrePost(
  RegData,
  reshID,
  outfile = "",
  hentData = 0,
  preprosess = 1,
  enhetsUtvalg = 1,
  valgtMaal = "Gjsn",
  valgtVar,
  datoFra = "2012-04-01",
  datoTil = "2050-12-31",
  datoFraSluttreg = "",
  datoTilSluttreg = "",
  minbmistart = 0,
  maxbmistart = 200,
  minald = 0,
  maxald = 130,
  erMann = "",
  regType = "",
  diagnose = ""
)
}
\arguments{
\item{valgtMaal}{Sentralmål 'Med' gir median, alt annet gir gjennomsnitt}

\item{valgtVar}{Variabelen det skal vises resultat for}

\item{diagnose}{}
}
\value{
Figur med gjennomsnitt eller median ved to ulike tidspunkt.
}
\description{
Pre/post-sammenligningsfigur har litt anndre datofiltre enn f.eks. figur for andelsfordelingsfigur. Dette fordi:
- datofilter for f.eks. andelsfordelingsfigur brukes i kombinasjon med filter for registreingstype
- mens datofilter for pre/post-figur ikke kan brukes i kombinasjon med registreringstypefilter,
 men må ta hensyn til at både start-og sluttregistrenger skal være med i figur.
Se ellers figurtekst som kommer ut til PrePost-figur i Shiny-applikasjonen
for å forstå hvilke datoer som er sortert på.
}
\examples{
/ ',
\%')
}
