% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NorSpis1Nokkeltall.R
\name{NorSpis1Nokkeltall}
\alias{NorSpis1Nokkeltall}
\title{Different key figures to display in dashboard}
\usage{
NorSpis1Nokkeltall(RegData, enhetsUtvalgEgenNasjonal = "nasjonal",
  reshID)
}
\arguments{
\item{RegData}{Registry data}

\item{enhetsUtvalgEgenNasjonal}{String user role}

\item{reshID}{String org id}
}
\value{
Ulike nøkkeltall:
- Antall startregistreringer
- Antall unike pasienter (ferdigstilte)
}
\description{
Different key figures to display in dashboard
}
