% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NorSpis1NokkeltallTid.R
\name{NorSpis1NokkeltallTid}
\alias{NorSpis1NokkeltallTid}
\title{Plot of key figures over time}
\usage{
NorSpis1NokkeltallTid(RegData, enhetsUtvalgEgenNasjonal = "nasjonal", reshID)
}
\arguments{
\item{RegData}{Registry data}

\item{enhetsUtvalgEgenNasjonal}{Whether key figures should show for own unit or be a national value}

\item{reshID}{}
}
\value{
ggplot2 figure og key figures over time:
- Antall startregistreringer (ferdigstilte)
}
\description{
Plot of key figures over time
}
