% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NorSpis1TabRegStatus.R
\name{NorSpis1TabRegStatus}
\alias{NorSpis1TabRegStatus}
\title{Table with number of registrations}
\usage{
NorSpis1TabRegStatus(RegData, userRole, reshID, datoFra, datoTil)
}
\arguments{
\item{RegData}{Registry data}

\item{datoTil}{}
}
\value{
A DT::datatable
}
\description{
Table with number of registrations within different registration types
at the different hospital units as well as the sum  nationally
}
