% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NorSpis1TabRegStatusUtvidet.R
\name{NorSpis1TabRegStatusUtvidet}
\alias{NorSpis1TabRegStatusUtvidet}
\title{Tabell med tellinger, per behandlingsenhet/sykehusavdeling}
\usage{
NorSpis1TabRegStatusUtvidet(
  RegData = RegData,
  userRole,
  reshID,
  datoFra,
  datoTil
)
}
\arguments{
\item{datoTil}{}
}
\value{
Tabell med tellinger, per behandlingsenhet/sykehusavdeling
- unike pasienter, per avdeling (ferdigstilte)
- startregistreringer (ferdigstilte)
- Manglende sluttregistreringer (til ferdigstilte startregistreringer)
- registrering startet men ikke fullført
}
\description{
Tabell med tellinger, per behandlingsenhet/sykehusavdeling
}
