% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_data.R
\name{query_data}
\alias{query_data}
\alias{query_alle_scorer}
\alias{query_behandling}
\alias{query_behandling_num}
\alias{query_enkelt_ledd}
\alias{query_enkelt_ledd_num}
\alias{forlops_oversikt}
\alias{query_forlops_oversikt}
\title{Get data from NorSpis}
\usage{
query_alle_scorer(registry_name, resh_id, ...)

query_behandling(registry_name, resh_id, ...)

query_behandling_num(registry_name, resh_id, ...)

query_enkelt_ledd(registry_name, resh_id, ...)

query_enkelt_ledd_num(registry_name, resh_id, ...)

query_forlops_oversikt(registry_name, resh_id, ...)
}
\arguments{
\item{registry_name}{Character string with name of the registry from which
data are to be queried}

\item{resh_id}{Integer providing the organization id. Usefull for data
filtering}

\item{...}{Optional arguments to pass to the function. If \code{session} is
provided this will be assumed a valid shiny session object and hence logging
will be performed}
}
\value{
Data frame of registry data
}
\description{
Functions to query data from a database holding NorSpis data at Rapporteket.
Providing these function with the shiny session object logging of queries
will also be performed
}
