% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_figFig_timetrend.R
\name{make_figFig_timetrend}
\alias{make_figFig_timetrend}
\title{Title}
\usage{
make_figFig_timetrend(variable01, add_fill = F)
}
\arguments{
\item{add_fill}{Add fill under line}
}
\value{

}
\description{
Title
}
