% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_figTable_unitCompar.R
\name{make_figTable_unitCompar}
\alias{make_figTable_unitCompar}
\title{Function to make data table for plot with comparison of hospitals
Making a small summary table which will go into the plot with  percentages (proportions) within each group (hospital):}
\usage{
make_figTable_unitCompar(myIndata_NatVal, myInvar01 = myVar01)
}
\arguments{
\item{myIndata_NatVal}{myIndata_NatVal <- RegData_NatVal}

\item{myInvar01}{myVar01 <- quo(PROP_PO10Pasientsikkerhet)

Note: This function's name when writing aarsrapport 2019 (in 2020) was: make_proptable_hospitals}
}
\value{
summary table with propoprtions at each unit/hosptital and one national value to go into make_figFig_unitCompar
}
\description{
Function to make data table for plot with comparison of hospitals
Making a small summary table which will go into the plot with  percentages (proportions) within each group (hospital):
}
