% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_table_popchar.R
\name{make_table_popchar}
\alias{make_table_popchar}
\title{make_table_popchar}
\usage{
make_table_popchar(RegData = myFilteredData, varsInTab = myvarString)
}
\arguments{
\item{varsInTab}{}
}
\value{

}
\description{
Function for table with overview of patient characteristics.
}
\details{
The name of this function when maing aarsrapport 2019, in 2020,  was make_popCharTab().
}
