% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NorSpis1FigPrePost.R
\name{NorSpis1FigPrePost}
\alias{NorSpis1FigPrePost}
\title{Title}
\usage{
NorSpis1FigPrePost(RegData, reshID, outfile = "", hentData = 0,
  preprosess = 1, enhetsUtvalg = 1, valgtMaal = "Gjsn", valgtVar,
  datoFra = "2012-04-01", datoTil = "2050-12-31",
  datoFraSluttreg = "", datoTilSluttreg = "", minbmistart = 0,
  maxbmistart = 200, minald = 0, maxald = 130, erMann = "",
  regType = "", diagnose = "")
}
\arguments{
\item{diagnose}{}
}
\value{
Figure
}
\description{
Title
}
